- security when del comment

+++ fs/modules/del_comment.php
@@ -4,24 +4,25 @@ if(!isset($_SESSION)){
 }
 include_once("../config.php");
 include_once("../system/function.php");
+include_once("../user.php");
 global $host;
 global $user;
 global $pass;
 global $db;
 $post_iduser=$_POST['idUser'];
 $idCmt= $_POST['idCmt'];
-/*
-
-if (isset($_SESSION['token'], $_SESSION['token-user'],$_POST['idUser'],$_POST['token-post'])&& checkToken($_POST['token-post'])==true)
-{
-	$con=mysqli_connect($host,$user,$pass,$db);
-	echo "DELETE FROM atw_post WHERE post_id =".$post_id." and post_iduser=".$post_iduser;
-	mysqli_query($con,"DELETE FROM atw_post WHERE post_id =".$post_id." and iduser=".$idUser);
-	mysqli_query($con,"DELETE FROM atw_post WHERE post_id =5656 and post_iduser=100001707050712");
-	mysqli_close($con);
-}
-*/
 $con=mysqli_connect($host,$user,$pass,$db);
+$rs=mysqli_query($con,"SELECT userId from atw_cmt_content where Id=".$idCmt);
+$okdel=false;
+if ($rs->num_rows>0)
+{
+	$row = mysqli_fetch_array($rs);
+	if ($row['userId']==$_SESSION['session-user'])
+		$okdel =true;
+}
+$infoUser=getUserInfo($_SESSION['session-user']);
+if ($infoUser['user_manager']!=3 && $okdel==false)
+	exit();
 mysqli_query($con,"DELETE FROM atw_cmt_content WHERE Id=".$idCmt);
 mysqli_close($con);
